Ended working with UserLogController.cs

+++ DiaDiary/Controllers/UserLogController.cs
@@ -12,29 +12,54 @@ public class UserLogController
     {
         options = new string[]
         {
-            "Create", "Read", "Update", "Delete", "Delete All", "About", "Exit"
+            "Create", "Read", "Update", "Delete", "Delete All", "Account", "About", "Exit"
         }
     };
 
     private readonly MongoRepository<UserLog> _mongoRepository;
     private readonly ApplicationUser _applicationUser;
+    private readonly ApplicationUserController _applicationUserController;
 
-    public UserLogController(MongoRepository<UserLog> mongoRepository, ApplicationUser applicationUser)
+    public UserLogController(MongoRepository<UserLog> mongoRepository,
+        MongoRepository<ApplicationUser> applicationUserMongoRepository, ApplicationUser applicationUser)
     {
         _mongoRepository = mongoRepository;
         _applicationUser = applicationUser;
+        _applicationUserController = new ApplicationUserController(applicationUserMongoRepository, _applicationUser);
     }
 
-    public dynamic ActionsRouting(int chosenAction) => chosenAction switch
+    public dynamic UserActions()
     {
-        0 => Create(_applicationUser),
-        1 => GetAll(_applicationUser),
-        2 => Update(_applicationUser),
-        3 => Delete(_applicationUser),
-        4 => DeleteAll(_applicationUser),
-        5 => ContextActions.About(),
-        6 => ContextActions.Exit(),
-    };
+        MenuElements menuElements =
+            new MenuElements()
+            {
+                options = new string[]
+                {
+                    "Create", "Read", "Update", "Delete", "Delete All", "Account", "About", "Exit"
+                }
+            };
+
+        dynamic ActionsRouting(int chosenAction) => chosenAction switch
+        {
+            0 => Create(_applicationUser),
+            1 => GetAll(_applicationUser),
+            2 => Update(_applicationUser),
+            3 => Delete(_applicationUser),
+            4 => DeleteAll(_applicationUser),
+            5 => _applicationUserController.AccountActions(_applicationUser),
+            6 => ContextActions.About(),
+            7 => ContextActions.Exit(),
+        };
+
+        while (_applicationUser.Email != null)
+        {
+            ScrollableMenu scrollableMenu = new ScrollableMenu(menuElements);
+            int chosenAction = scrollableMenu.Run();
+            dynamic execution = ActionsRouting(chosenAction);
+        }
+
+        return _applicationUser;
+    }
 
     private async Task Create(ApplicationUser applicationUser)
     {
@@ -53,14 +78,14 @@ public class UserLogController
     private async Task Update(ApplicationUser applicationUser)
     {
         Clear();
-        WriteLine("Enter Id of log you want to replace");
+        Messages.IdOfLogToReplace();
         string? idToUpdate = Console.ReadLine();
         var userLogToDelete = _mongoRepository.GetOne(x => x.Id == idToUpdate).Result;
         if (userLogToDelete.AssignedTo.Id == applicationUser.Id)
         {
             UserLog updatedUserLog = UserLogView.Update(idToUpdate);
             updatedUserLog.AssignedTo = applicationUser;
-            
+
             await _mongoRepository.Update(x => x.Id == idToUpdate, updatedUserLog);
         }
         else
@@ -89,6 +114,18 @@ public class UserLogController
 
     private async Task DeleteAll(ApplicationUser applicationUser)
     {
-        await _mongoRepository.DeleteRange(_ => _.AssignedTo == applicationUser);
+        Clear();
+        Messages.WarningToDeleteAll();
+        ConsoleKey keyPressed = ReadKey().Key;
+        if (keyPressed == ConsoleKey.Enter)
+        {
+            Messages.SuccessfulDeletion();
+            await _mongoRepository.DeleteRange(_ => _.AssignedTo == applicationUser);
+        }
+
+        if (keyPressed == ConsoleKey.Backspace)
+        {
+            return;
+        }
     }
 }
